+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### [Zanabazar](https://en.wikipedia.org/wiki/Zanabazar_junior)
 
+### 1.1743
+
+* Fixed listing personal root namespace
+* Small improvements and minor bug fixes
+* Updated translations
+
 ### 1.1742
 
 * Small improvements and minor bug fixes